requirements.txt
aioredis==2.0.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
bcrypt==4.1.2
black==24.2.0
cffi==1.16.0
click==8.1.7
cryptography==42.0.5
dnspython==2.6.1
email_validator==2.1.1
fastapi==0.110.0
h11==0.14.0
idna==3.6
motor==3.3.2
mypy-extensions==1.0.0
packaging==23.2
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.2.0
pycparser==2.21
pydantic==2.6.3
pydantic_core==2.16.3
PyJWT==2.8.0
pymongo==4.6.2
python-dotenv==1.0.1
redis==5.0.3
sniffio==1.3.1
starlette==0.36.3
typing_extensions==4.10.0
uvicorn==0.27.1